fix: global bug with replacing Menu in memory
See the next function calling:
-> Use Screen 1
-> User press key
 -> Call menu actions
   -> Some action change screen
     -> Screen 1 terminated
     -> Screen 2 Initialized

As you see, when the screen changes, the screen menu removes the screen in which he is. And this action happened in the static method.

When "Screen 1 terminated", the menu from Screen 1 deleting. It is good. But, after "Screen 2 Initialized" the menu from Screen 2 have the same address as the menu from Screen 1.

And because we are still in a loop (static method), the loop does not stop. And if the menu from Screen 2 exists an action with a key from screen 1, it will be called.

To prevent it, I move the function of init a screen before deleting the previous one. It does not solve problems with nested function calling, but the menu from screen 2 will have another address.

(nested calling will be removed in the next commit)