refactor: remove nested calls in ScreenManager
From the previous commit, ScreenManager has a problem with nested calling. For removing it, they will be changed the Screen class and add a new concept with the Screen signal.

The screen has no access to his ScreenManager, and when the active screen wants to change Screen it will store the screen "change screen" signal.

How it works:
-> Screen 1 run by manager
-> User press key
  -> Call menu actions
    -> Some action changes the active screen
      -> Screen 1 changing his signal
  -> Call menu actions ended
-> Screen 1 run end
-> ScreenManager looks at the signal Screen 1
  -> ScreenManager changes screen

As you see, nested calling will be removed. The screen run method will be called and ended.